+++ MANIFEST
@@ -1,6 +1,6 @@
 # file GENERATED by distutils, do NOT edit
 setup.cfg
 setup.py
-mou_translator/__init__.py
-mou_translator/__main__.py
-mou_translator/core.py
+mtranslate/__init__.py
+mtranslate/__main__.py
+mtranslate/core.py
