+++ README.md
@@ -7,4 +7,4 @@ as simple as that:
 translate("Bonjour","en","auto") will return "Hello"
 
 Also on PyPi:
-https://pypi.python.org/pypi?:action=display&name=mou_translator&version=1.1
+https://pypi.python.org/pypi?:action=display&name=mtranslate&version=1.2
