+++ mtranslate/core.py
@@ -1,0 +1,76 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+MIT License
+
+Copyright (c) 2016 Arnaud Aliès
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+import sys
+import re
+
+if (sys.version_info[0] < 3):
+    import urllib2
+    import urllib
+else:
+    import urllib.request
+    import urllib.parse
+
+agent = {'User-Agent':
+    "Mozilla/4.0 (\
+    compatible;\
+    MSIE 6.0;\
+    Windows NT 5.1;\
+    SV1;\
+    .NET CLR 1.1.4322;\
+    .NET CLR 2.0.50727;\
+    .NET CLR 3.0.04506.30\
+    )"}
+
+
+def translate(to_translate, to_language="auto", language="auto"):
+    """
+    Returns the translation using google translate
+    you must shortcut the language you define
+    (French = fr, English = en, Spanish = es, etc...)
+    if not defined it will detect it or use english by default
+
+    Example:
+    print(translate("salut tu vas bien?", "en"))
+    hello you alright?
+    """
+    base_link = "http://translate.google.com/m?hl=%s&sl=%s&q=%s"
+    if (sys.version_info[0] < 3):
+        to_translate = urllib.pathname2url(to_translate)
+        link = base_link % (to_language, language, to_translate)
+        request = urllib2.Request(link, headers=agent)
+        page = urllib2.urlopen(request).read()
+    else:
+        to_translate = urllib.parse.quote(to_translate)
+        link = base_link % (to_language, language, to_translate)
+        request = urllib.request.Request(link, headers=agent)
+        page = urllib.request.urlopen(request).read().decode("utf-8")
+    expr = r'class="t0">(.*?)<'
+    result = re.findall(expr, page)
+    if (len(result) == 0):
+        return ("")
+    return(result[0])
